Fix broken LinkedIn link in Github README.md
This change fixes a broken LinkedIn link in the Github README.md file. The previous link was not functioning correctly, causing a dead end for users who wanted to learn more about the person. The link has now been updated and tested to ensure that it redirects to the correct LinkedIn profile page.